Int type

+++ cpp/dolfinx/graph/partition.cpp
@@ -352,7 +352,7 @@ graph::build::distribute(MPI_Comm comm, std::span<const std::int64_t> list,
   std::vector<int> ghost_index_owner;
   std::vector<std::int64_t> global_indices, global_indices1;
   std::vector<int> src_ranks, src_ranks1;
-  for (std::int32_t p = 0; p < recv_disp.size() - 1; ++p)
+  for (std::size_t p = 0; p < recv_disp.size() - 1; ++p)
   {
     int src_rank = src[p];
     for (std::int32_t q = recv_disp[p]; q < recv_disp[p + 1]; ++q)
